added key bindings for insert cell; added rel_cells and added options
